feat(ReservationTimeService): 특정 시간에 대한 예약이 존재하는데, 그 시간을 삭제하려 할 때 예외처리
- ReservationRepository의 findAllByTimeId 추가

+++ src/main/java/roomescape/reservationtime/service/ReservationTimeService.java
@@ -3,6 +3,8 @@ package roomescape.reservationtime.service;
 import java.util.List;
 import java.util.NoSuchElementException;
 import org.springframework.stereotype.Service;
+import roomescape.reservation.model.Reservation;
+import roomescape.reservation.repository.ReservationRepository;
 import roomescape.reservationtime.dto.request.CreateReservationTimeRequest;
 import roomescape.reservationtime.dto.response.CreateReservationTimeResponse;
 import roomescape.reservationtime.dto.response.FindReservationTimeResponse;
@@ -12,12 +14,17 @@ import roomescape.reservationtime.repository.ReservationTimeRepository;
 @Service
 public class ReservationTimeService {
     private final ReservationTimeRepository reservationTimeRepository;
+    private final ReservationRepository reservationRepository;
 
-    public ReservationTimeService(final ReservationTimeRepository reservationTimeRepository) {
+    public ReservationTimeService(
+            final ReservationTimeRepository reservationTimeRepository,
+            final ReservationRepository reservationRepository) {
         this.reservationTimeRepository = reservationTimeRepository;
+        this.reservationRepository = reservationRepository;
     }
 
-    public CreateReservationTimeResponse createReservationTime(final CreateReservationTimeRequest createReservationTimeRequest) {
+    public CreateReservationTimeResponse createReservationTime(
+            final CreateReservationTimeRequest createReservationTimeRequest) {
         Long id = reservationTimeRepository.save(createReservationTimeRequest.toReservationTime());
 
         // TODO: 수정
@@ -40,6 +47,12 @@ public class ReservationTimeService {
     public void deleteById(final Long id) {
         reservationTimeRepository.findById(id)
                 .orElseThrow(() -> new NoSuchElementException("해당하는 예약 시간이 존재하지 않습니다."));
+        List<Reservation> reservations = reservationRepository.findAllByTimeId(id);
+        if (!reservations.isEmpty()) {
+            // TODO: 적절한 예외 던지기
+            throw new IllegalStateException("시간을 사용 중인 예약이 존재합니다.");
+        }
+
         reservationTimeRepository.deleteById(id);
     }
 }
